# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "BatchCreate": {
        "properties": {
          "filename": {
            "maxLength": 180,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 10000,
            "minItems": 1,
            "title": "Records",
            "type": "array"
          }
        },
        "required": [
          "filename",
          "records"
        ],
        "title": "BatchCreate",
        "type": "object"
      },
...
      "BatchStageResponse": {
        "properties": {
          "batch": {
            "$ref": "#/components/schemas/BatchSummary"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "sample_fields": {
            "items": {
              "type": "string"
            },
            "title": "Sample Fields",
            "type": "array"
          }
        },
        "required": [
          "batch",
          "sample_fields",
          "message"
        ],
        "title": "BatchStageResponse",
        "type": "object"
      },
      "ImportCreate": {
        "properties": {
          "dataset_id": {"type": "string"},
          "filename": {"maxLength": 180, "minLength": 1, "type": "string"},
          "media_type": {"type": "string"},
          "size_bytes": {"minimum": 1, "type": "integer"}
        },
        "required": ["dataset_id", "filename", "size_bytes"],
        "type": "object"
      },
      "ImportResponse": {
        "properties": {
          "import_id": {"type": "string"},
          "upload_id": {"type": "string"},
          "upload_url": {"type": "string"},
          "state": {"enum": ["created", "uploading", "profiling", "queued", "running", "needs_input", "succeeded", "failed", "canceled"], "type": "string"},
          "offset": {"minimum": 0, "type": "integer"},
          "size_bytes": {"minimum": 1, "type": "integer"},
          "format": {"type": "string"},
          "job_id": {"type": "string"}
        },
        "required": ["import_id", "upload_id", "upload_url", "state", "offset", "size_bytes"],
        "type": "object"
      },
      "JobResponse": {
        "properties": {
          "job_id": {"type": "string"},
          "import_id": {"type": "string"},
          "run_id": {"type": "string"},
          "state": {"enum": ["queued", "running", "needs_input", "succeeded", "failed", "canceled"], "type": "string"},
          "status_url": {"type": "string"},
          "events_url": {"type": "string"},
          "error": {"type": "string"}
        },
        "required": ["job_id", "import_id", "state", "status_url", "events_url"],
        "type": "object"
      },
      "StagePage": {
        "properties": {
          "stage_id": {"type": "string"},
          "rows": {"items": {"additionalProperties": true, "type": "object"}, "type": "array"},
          "next_cursor": {"type": "string"},
          "returned": {"minimum": 0, "type": "integer"}
        },
        "required": ["stage_id", "rows", "returned"],
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "All-Go local-first, white-box data curation with optional ingestion and durable orchestration adapters.",
    "title": "Verity API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/imports": {
      "post": {
        "operationId": "create_import",
        "parameters": [{"in": "header", "name": "Idempotency-Key", "required": true, "schema": {"maxLength": 200, "type": "string"}}],
        "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImportCreate"}}}, "required": true},
        "responses": {"202": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImportResponse"}}}, "description": "Resumable import created"}},
        "summary": "Create Import"
      }
    },
    "/api/v1/uploads/{upload_id}": {
      "head": {
        "operationId": "inspect_upload",
        "parameters": [{"in": "path", "name": "upload_id", "required": true, "schema": {"type": "string"}}],
        "responses": {"204": {"description": "Current upload offset", "headers": {"Upload-Offset": {"schema": {"type": "integer"}}}}},
        "summary": "Inspect Upload Offset"
      },
      "patch": {
        "operationId": "append_upload",
        "parameters": [
          {"in": "path", "name": "upload_id", "required": true, "schema": {"type": "string"}},
          {"in": "header", "name": "Upload-Offset", "required": true, "schema": {"minimum": 0, "type": "integer"}}
        ],
        "requestBody": {"content": {"application/offset+octet-stream": {"schema": {"format": "binary", "type": "string"}}}, "required": true},
        "responses": {"204": {"description": "Upload chunk committed"}},
        "summary": "Append Upload Chunk"
      }
    },
    "/api/v1/imports/{import_id}/complete": {
      "post": {
        "operationId": "complete_import",
        "parameters": [
          {"in": "path", "name": "import_id", "required": true, "schema": {"type": "string"}},
          {"in": "header", "name": "Idempotency-Key", "required": true, "schema": {"maxLength": 200, "type": "string"}}
        ],
        "responses": {"202": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobResponse"}}}, "description": "Automatic pipeline job created"}},
        "summary": "Complete Import And Start Pipeline"
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job",
        "parameters": [{"in": "path", "name": "job_id", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/JobResponse"}}}, "description": "Job status"}},
        "summary": "Get Job"
      }
    },
    "/api/v1/jobs/{job_id}/events": {
      "get": {
        "operationId": "stream_job_events",
        "parameters": [{"in": "path", "name": "job_id", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"content": {"text/event-stream": {"schema": {"type": "string"}}}, "description": "Replayable job event stream"}},
        "summary": "Stream Job Events"
      }
    },
    "/api/v1/stages/{stage_id}/records": {
      "get": {
        "operationId": "page_stage_records",
        "parameters": [
          {"in": "path", "name": "stage_id", "required": true, "schema": {"type": "string"}},
          {"in": "query", "name": "cursor", "required": false, "schema": {"type": "string"}},
          {"in": "query", "name": "limit", "required": false, "schema": {"default": 50, "maximum": 200, "minimum": 1, "type": "integer"}}
        ],
        "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/StagePage"}}}, "description": "Bounded page of stage records"}},
        "summary": "Page Stage Records"
      }
    },
    "/api/v1/outputs/{output_id}": {
      "get": {
        "operationId": "download_output",
        "parameters": [{"in": "path", "name": "output_id", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"content": {"application/octet-stream": {"schema": {"format": "binary", "type": "string"}}}, "description": "Output artifact", "headers": {"X-Verity-SHA256": {"description": "SHA-256 digest of the downloaded artifact", "schema": {"type": "string"}}}}},
        "summary": "Download Output"
      }
    },
    "/api/v1/datasets/{dataset_id}/batches": {
      "post": {
        "operationId": "stage_batch_api_v1_datasets__dataset_id__batches_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "description": "Stable key used to deduplicate batch staging retries.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStageResponse"
                }
              }
            },
            "description": "Idempotent retry of an existing staged batch"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stage Batch"
      }
    },
    "/api/v1/integrations/airbyte/jobs/{job_id}": {
      "get": {
        "operationId": "get_airbyte_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {"minimum": 1, "type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AirbyteJob"}}},
            "description": "Airbyte job status"
          },
          "503": {"description": "Airbyte integration is not configured"}
        },
        "summary": "Get Airbyte Job"
      }
    },
    "/api/v1/integrations/airbyte/syncs": {
      "post": {
        "operationId": "trigger_airbyte_sync",
        "requestBody": {
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AirbyteSyncRequest"}}},
          "required": true
        },
        "responses": {
          "202": {
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AirbyteJob"}}},
            "description": "Airbyte sync accepted"
          },
          "503": {"description": "Airbyte integration is not configured"}
        },
        "summary": "Trigger Airbyte Sync"
      }
    },
...
    "/api/v1/reviews/{record_id}": {
      "patch": {
        "operationId": "update_review_api_v1_reviews__record_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Review"
      }
    },
...
    "/api/v1/stages/{stage_id}/comparison": {
      "get": {
        "operationId": "stage_comparison",
        "parameters": [
          {
            "in": "path",
            "name": "stage_id",
            "required": true,
            "schema": {"type": "string"}
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {"default": 8, "maximum": 20, "minimum": 1, "type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "content": {"application/json": {"schema": {"additionalProperties": true, "type": "object"}}},
            "description": "Stage input, output, and representative record comparisons"
          },
          "422": {
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}},
            "description": "Validation Error"
          }
        },
        "summary": "Compare Stage Boundary"
      }
    },
    "/api/v1/stages/{stage_id}/preview": {
      "get": {
        "operationId": "stage_preview_api_v1_stages__stage_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "stage_id",
            "required": true,
            "schema": {
              "title": "Stage Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Stage Preview Api V1 Stages  Stage Id  Preview Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stage Preview"
      }
    },
...
  }
}